model/requirements.txt
pandas==1.4.2
numpy==1.22.3
scikit-learn==1.0.2
joblib==1.1.0
lightgbm==3.3.2
xgboost==1.6.1
redis==4.1.4
pytest==7.1.1
python-dotenv==0.21.0 